Read files from directories for x509_cert plugin
The x509_cert input plugin before, could read only files to which the
complete path has been provided in the sources configuration. This
enhancement allows the plugin to parse a path provided to a directory
and read all files from it. Additionally, support for wild card
patterns in file path name has also been added. This allows the plugin
to parse all the files matching the provided pattern and read from them
respectively.